+++ src/integrations/crm/MicrosoftDynamics365.php
@@ -14,7 +14,6 @@ use Craft;
 use craft\helpers\App;
 use craft\helpers\ArrayHelper;
 use craft\helpers\Json;
-
 use League\OAuth1\Client\Server\Server as Oauth1Provider;
 use League\OAuth2\Client\Provider\AbstractProvider;
 use TheNetworg\OAuth2\Client\Provider\Azure;
@@ -60,10 +59,12 @@ class MicrosoftDynamics365 extends Crm
     public bool $mapToLead = false;
     public bool $mapToOpportunity = false;
     public bool $mapToAccount = false;
+    public bool $mapToIncident = false;
     public ?array $contactFieldMapping = null;
     public ?array $leadFieldMapping = null;
     public ?array $opportunityFieldMapping = null;
     public ?array $accountFieldMapping = null;
+    public ?array $incidentFieldMapping = null;
 
     private array $_entityOptions = [];
 
@@ -123,6 +124,7 @@ class MicrosoftDynamics365 extends Crm
         $lead = $this->getFormSettingValue('lead');
         $opportunity = $this->getFormSettingValue('opportunity');
         $account = $this->getFormSettingValue('account');
+        $incident = $this->getFormSettingValue('incident');
 
         // Validate the following when saving form settings
         $rules[] = [
@@ -149,6 +151,12 @@ class MicrosoftDynamics365 extends Crm
             }, 'on' => [Integration::SCENARIO_FORM],
         ];
 
+        $rules[] = [
+            ['incidentFieldMapping'], 'validateFieldMapping', 'params' => $incident, 'when' => function($model) {
+                return $model->enabled && $model->mapToIncident;
+            }, 'on' => [Integration::SCENARIO_FORM]
+        ];
+
         return $rules;
     }
 
@@ -161,12 +169,14 @@ class MicrosoftDynamics365 extends Crm
             $leadFields = $this->_getEntityFields('lead');
             $opportunityFields = $this->_getEntityFields('opportunity');
             $accountFields = $this->_getEntityFields('account');
+            $incidentFields = $this->_getEntityFields('incident');
 
             $settings = [
                 'contact' => $contactFields,
                 'lead' => $leadFields,
                 'opportunity' => $opportunityFields,
                 'account' => $accountFields,
+                'incident' => $incidentFields,
             ];
         } catch (Throwable $e) {
             Integration::apiError($this, $e);
@@ -182,11 +192,13 @@ class MicrosoftDynamics365 extends Crm
             $leadValues = $this->getFieldMappingValues($submission, $this->leadFieldMapping, 'lead');
             $opportunityValues = $this->getFieldMappingValues($submission, $this->opportunityFieldMapping, 'opportunity');
             $accountValues = $this->getFieldMappingValues($submission, $this->accountFieldMapping, 'account');
+            $incidentValues = $this->getFieldMappingValues($submission, $this->incidentFieldMapping, 'incident');
 
             $contactId = null;
             $leadId = null;
             $opportunityId = null;
             $accountId = null;
+            $incidentId = null;
 
             if ($this->mapToContact) {
                 $contactPayload = $contactValues;
@@ -283,6 +295,31 @@ class MicrosoftDynamics365 extends Crm
                     Integration::error($this, Craft::t('formie', 'Missing return opportunityid {response}. Sent payload {payload}', [
                         'response' => Json::encode($response),
                         'payload' => Json::encode($opportunityPayload),
+                    ]), true);
+
+                    return false;
+                }
+            }
+
+            if ($this->mapToIncident) {
+                $incidentPayload = $incidentValues;
+
+                if ($contactId) {
+                    $incidentPayload['customerid_contact@odata.bind'] = $this->_formatLookupValue('contacts', $contactId);
+                }
+
+                $response = $this->deliverPayload($submission, 'incidents?$select=incidentid', $incidentPayload);
+
+                if ($response === false) {
+                    return true;
+                }
+
+                $incidentId = $response['incidentid'] ?? '';
+
+                if (!$incidentId) {
+                    Integration::error($this, Craft::t('formie', 'Missing return incidentid {response}. Sent payload {payload}', [
+                        'response' => Json::encode($response),
+                        'payload' => Json::encode($incidentPayload),
                     ]), true);
 
                     return false;
@@ -602,6 +639,11 @@ class MicrosoftDynamics365 extends Crm
                 'entity' => 'leads',
                 'label' => 'fullname',
                 'value' => 'leadid',
+            ],
+            'incident' => [
+                'entity' => 'incidents',
+                'label' => 'title',
+                'value' => 'incidentid',
             ],
             'transactioncurrency' => [
                 'entity' => 'transactioncurrencies',